Move aria label patches back into base.ts
